+++ SLM/chadgpt_sft.py
@@ -50,8 +50,8 @@ def enable_position_aware_rope(pi_scale=1.0):
             t = torch.arange(start_pos, start_pos + seq_len, device=x.device).unsqueeze(1).float()
             t = t * pi_scale
         freqs = (t * self.theta).unsqueeze(0).unsqueeze(0)
-        cos = torch.cos(freqs).repeat_interleave(2, dim=-1)
-        sin = torch.sin(freqs).repeat_interleave(2, dim=-1)
+        cos = torch.cos(freqs).repeat_interleave(2, dim=-1).to(dtype=x.dtype)
+        sin = torch.sin(freqs).repeat_interleave(2, dim=-1).to(dtype=x.dtype)
         return x * cos + self._rotate(x) * sin
 
     RoPEEmbedding.forward = _rope_forward
